% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{access_cache}
\alias{access_cache}
\title{Load cached script execution results.}
\usage{
access_cache(
  cache_path,
  script_path,
  message = "Loading cached results",
  envir = globalenv(),
  crash = TRUE,
  encoding = "UTF-8"
)
}
\arguments{
\item{cache_path}{a character string specifying the path and name of the
cached file to be loaded (usually \code{.RData} or \code{.RDA}).}

\item{script_path}{a character string specifying the path to the R script to
be executed if the cached file does not exist.}

\item{message}{user defined text to be displayed prior to loading of
the cached file.}

\item{envir}{the environment where the cached file data should be loaded.}

\item{crash}{logical, should the execution of the R script be halted on error
and the error returned?}

\item{encoding}{encoding of the script, UTF-8 per default.}
}
\description{
A short description...
}
