% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_functions.R
\name{enter_directory}
\alias{enter_directory}
\title{Go to or create a directory}
\usage{
enter_directory(path, create = FALSE, showWarnings = TRUE, ...)
}
\arguments{
\item{path}{a character vector containing a single path name.}

\item{create}{logical; should the directory be created if it does not exist yet?}

\item{showWarnings}{logical; should the warnings on failure be shown?}

\item{...}{other arguments passed to dir.create().}
}
\value{
Nothing, called for its side effects.
}
\description{
A handy wrapper around the base functions dir.exists and dir.create.
}
\details{
If 'create' is set to FALSE, the function raises an error, if the path does not exist.
}
