% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_functions.R
\name{go_proj_directory}
\alias{go_proj_directory}
\title{Go back to the project directory}
\usage{
go_proj_directory()
}
\value{
Nothing, called for its side effects.
}
\description{
Sets the working directory back to the RStudio project directory.
}
