% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_functions.R
\name{go_up_directory}
\alias{go_up_directory}
\title{Go up in the directory tree}
\usage{
go_up_directory()
}
\value{
Nothing, called for its side effects.
}
\description{
Moves up in the directory tree relative to the current working directory.
}
