% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_functions.R
\name{insert_head}
\alias{insert_head}
\title{Insert a header with the name of the sourced script}
\usage{
insert_head(prefix = "Executing:")
}
\arguments{
\item{prefix}{a user-specified prefix displayed before the script name.}
}
\value{
Nothing, called for its side effects.
}
\description{
Displays a message with the name of the sourced script.
}
