% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_functions.R
\name{insert_msg}
\alias{insert_msg}
\title{Insert a message displayed during file sourcing}
\usage{
insert_msg(text = NULL, separator = ">", sep_len = 60)
}
\arguments{
\item{text}{user-specified message text.}

\item{separator}{separator sign.}

\item{sep_len}{separator length.}
}
\value{
Nothing, called for its side effects.
}
\description{
Displays a user defined message during sourcing of the script.
}
