% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_functions.R
\name{insert_tail}
\alias{insert_tail}
\title{Insert a trailer with the name of the sourced script}
\usage{
insert_tail(def_suffix = "succesfully sourced")
}
\arguments{
\item{prefix}{a user-specified prefix displayed before the script name.}
}
\value{
Nothing, called for its side effects.
}
\description{
Displays a message with the name of the sourced script.
and user-specified confirmation of the successful execution.
}
