% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_functions.R
\name{list_file_type}
\alias{list_file_type}
\title{List files with the given extension}
\usage{
list_file_type(path = ".", extension = ".txt")
}
\arguments{
\item{path}{a character vector containing a single path name.}

\item{extension}{file extension.}
}
\value{
Nothing, called for its side effects.
}
\description{
Lists all files in the given directory with the given extension.
}
