% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{source_all}
\alias{source_all}
\title{Source safely a pipeline of R files}
\usage{
source_all(paths, encoding = "UTF-8", message = FALSE, crash = FALSE, ...)
}
\arguments{
\item{paths}{a vector with the file paths.}

\item{encoding}{encoding of the script, UTF-8 per default.}

\item{message}{logical, should a message with the file name be displayed prior to execution?}

\item{crash}{logical, should an error be returned? Useful during debugging.}

\item{...}{additional arguments passed to source().}
}
\value{
a data frame with the script names, execution status and possible errors.
}
\description{
Source safely a pipeline of R files
}
\details{
sources the scrips, the first error per script is captured. Useful in multi-file analysis pipelines:
if 'crash' is FALSE, single execution errors are captured but do not necessarily crash the pipeline.
}
