% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{source_safe}
\alias{source_safe}
\alias{source_all}
\title{Source safely R files}
\usage{
source_safe(path, encoding = "UTF-8", message = FALSE, crash = FALSE, ...)

source_all(paths, encoding = "UTF-8", message = FALSE, crash = FALSE, ...)
}
\arguments{
\item{path}{path to an R file.}

\item{encoding}{encoding of the script, UTF-8 per default.}

\item{message}{logical, should a message with the file name be displayed
prior to and after execution?}

\item{crash}{logical, should an error be returned? Useful during debugging.}

\item{...}{additional arguments passed to \code{source()}.}

\item{paths}{paths to R files.}
}
\value{
a data frame with the script name, execution status and
possible errors.
}
\description{
\code{source_safe()} and \code{source_all()} source respectively one or more R scripts
and return their execution status. The first error is captured and, if
\code{crash = TRUE}, halts the execution of the script pipeline.
#'
}
