% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_functions.R
\name{insert_msg}
\alias{insert_msg}
\alias{insert_head}
\alias{insert_tail}
\title{Messages displayed during R file sourcing}
\usage{
insert_msg(text = NULL, separator = ">", sep_len = 60)

insert_head(prefix = "Executing:")

insert_tail(suffix = "succesfully sourced")
}
\arguments{
\item{text}{user-specified message text.}

\item{separator}{separator sign.}

\item{sep_len}{separator length.}

\item{prefix}{a user-specified text displayed before the script name.}

\item{suffix}{a user-specified text displayed after the script name.}
}
\value{
Nothing, called for its side effects.
}
\description{
\code{insert_msg()} displays a user defined message during sourcing of the script.
\code{insert_head()} prints a message with the name of the sourced script.
\code{insert_tail()} displays a message with the name of the sourced script and
user-specified confirmation of the successful execution.
}
