% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{source_safe}
\alias{source_safe}
\title{Source safely the R file}
\usage{
source_safe(path, encoding = "UTF-8", message = FALSE, crash = FALSE, ...)
}
\arguments{
\item{path}{path to the R file.}

\item{encoding}{encoding of the script, UTF-8 per default.}

\item{message}{logical, should a message with the file name be displayed prior to execution?}

\item{crash}{logical, should an error be returned? Useful during debugging.}

\item{...}{additional arguments passed to source().}
}
\value{
a data frame with the script name, execution status and possible errors.
}
\description{
Source safely the R file
}
\details{
sources the script, the first error is captured. Useful in multi-file analysis pipelines.
}
